Tidy up tests, separate them into files for more obvious organization
